core: big refactor and move stuff around
